Refactor asset action execution helpers
Manual smoke: run createInstantHustle action via node script to verify payout, time spend, and log entry.
